Adding a footer with the repository linked
In the App.js edited the function Apps return where I added a footer
in which I included two links
to access information about me and this repository project.
Link to my portfolio ("Simona K... H...") and GitHub documentation link ("open-sourced on GitHub").

Using the App.js as the main component for this application.